py/pyproject.toml
[tool.poetry]
name = "gooey-gui"
version = "0.1.0"
description = ""
authors = ["Dev Aggarwal <devxpy@gmail.com>"]
license = "FSL-1.1-Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^1.10.12"
fastapi = "^0.85.2"
redis = "^4.5.1"
uvicorn = { extras = ["standard"], version = "^0.18.3" }
furl = "^2.1.3"
python-multipart = "^0.0.6"
python-decouple = "^3.6"
loguru = "^0.7.2"

opencv-contrib-python = { version = "^4.7.0.72", optional = true }
numpy = { version = "^1.25.0", optional = true }

[tool.poetry.extras]
image = ["opencv-contrib-python", "numpy"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
